requirements.txt
crewai
google-generativeai
python-dotenv
streamlit
jinja2
pydantic
python-docx